+++ package.json
@@ -106,8 +106,7 @@
     "src/themes/*",
     "src/presets/*",
     "src/tools/*",
-    "src/website",
-    "src/utils"
+    "src/website"
   ],
   "husky": {
     "hooks": {
