+++ src/implementations/vanilla/components/table/package.json
@@ -10,7 +10,7 @@
   "module": "table.js",
   "dependencies": {
     "@ecl/theme-dev": "^3.0.0",
-    "@ecl/utils": "^3.0.0"
+    "@ecl/dom-utils": "^3.0.0"
   },
   "private": true,
   "repository": {
