+++ src/presets/dev/package.json
@@ -8,7 +8,7 @@
   "sass": "src/dev.scss",
   "private": true,
   "dependencies": {
-    "@ecl/utils": "^3.0.0",
+    "@ecl/dom-utils": "^3.0.0",
     "@ecl/vanilla-component-blockquote": "^3.0.0",
     "@ecl/vanilla-component-button": "^3.0.0",
     "@ecl/vanilla-component-checkbox": "^3.0.0",
