+++ src/tools/dom-utils/autoinit/README.md
