+++ src/tools/dom-utils/autoinit/index.js
