+++ src/tools/dom-utils/package.json
@@ -1,9 +1,9 @@
 {
-  "name": "@ecl/utils",
+  "name": "@ecl/dom-utils",
   "author": "European Commission",
   "license": "EUPL-1.1",
   "version": "3.0.0",
-  "description": "ECL utilities",
+  "description": "ECL DOM utilities",
   "main": "index.js",
   "module": "index.js",
   "private": true,
